@@ -3,10 +3,13 @@
 
 using k8s;
 using k8s.Models;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Microsoft.Kubernetes;
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Linq;
 using Yarp.Kubernetes.Controller.Services;
 
 namespace Yarp.Kubernetes.Controller.Caching;
@@ -18,16 +21,79 @@ namespace Yarp.Kubernetes.Controller.Caching;
 public class IngressCache : ICache
 {
     private readonly object _sync = new object();
+    private readonly Dictionary<string, IngressClassData> _ingressClassData = new Dictionary<string, IngressClassData>();
     private readonly Dictionary<string, NamespaceCache> _namespaceCaches = new Dictionary<string, NamespaceCache>();
+    private readonly YarpOptions _options;
+    private readonly ILogger<IngressCache> _logger;
 
-    public void Update(WatchEventType eventType, V1Ingress ingress)
+    private bool _isDefaultController;
+
+    public IngressCache(IOptions<YarpOptions> options, ILogger<IngressCache> logger)
+    {
+        _options = options?.Value ?? throw new ArgumentNullException(nameof(options));
+        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+    }
+
+    public void Update(WatchEventType eventType, V1IngressClass ingressClass)
+    {
+        if (ingressClass is null)
+        {
+            throw new ArgumentNullException(nameof(ingressClass));
+        }
+
+        if (!string.Equals(_options.ControllerClass, ingressClass.Spec.Controller, StringComparison.OrdinalIgnoreCase))
+        {
+#pragma warning disable CA1303 // Do not pass literals as localized parameters
+            _logger.LogInformation(
+                "Ignoring {IngressClassNamespace}/{IngressClassName} as the spec.controller is not the same as this ingress",
+                ingressClass.Metadata.NamespaceProperty,
+                ingressClass.Metadata.Name);
+#pragma warning restore CA1303 // Do not pass literals as localized parameters
+            return;
+        }
+
+        var ingressClassName = ingressClass.Name();
+        lock (_sync)
+        {
+            if (eventType == WatchEventType.Added || eventType == WatchEventType.Modified)
+            {
+                _ingressClassData[ingressClassName] = new IngressClassData(ingressClass);
+            }
+            else if (eventType == WatchEventType.Deleted)
+            {
+                _ingressClassData.Remove(ingressClassName);
+            }
+
+            _isDefaultController = _ingressClassData.Values.Any(ic => ic.IsDefault);
+        }
+    }
+
+    public bool Update(WatchEventType eventType, V1Ingress ingress)
     {
         if (ingress is null)
         {
             throw new ArgumentNullException(nameof(ingress));
         }
 
-        Namespace(ingress.Namespace()).Update(eventType, ingress);
+        if (IsYarpIngress(ingress.Spec))
+        {
+            Namespace(ingress.Namespace()).Update(eventType, ingress);
+            return true;
+        }
+
+#pragma warning disable CA1303 // Do not pass literals as localized parameters
+        if (eventType == WatchEventType.Modified && Namespace(ingress.Namespace()).IngressExists(ingress))
+        {
+            // Special handling for an ingress that has the ingressClassName removed
+            _logger.LogInformation("Removing ingress {IngressNamespace}/{IngressName} because of unknown ingress class", ingress.Metadata.NamespaceProperty, ingress.Metadata.Name);
+            Namespace(ingress.Namespace()).Update(WatchEventType.Deleted, ingress);
+            return true;
+        }
+
+        _logger.LogInformation("Ignoring ingress {IngressNamespace}/{IngressName} because of ingress class", ingress.Metadata.NamespaceProperty, ingress.Metadata.Name);
+#pragma warning restore CA1303 // Do not pass literals as localized parameters
+
+        return false;
     }
 
 
@@ -75,6 +141,19 @@ public class IngressCache : ICache
         }
 
         return ingresses;
+    }
+
+    private bool IsYarpIngress(V1IngressSpec spec)
+    {
+        if (spec.IngressClassName != null)
+        {
+            lock (_sync)
+            {
+                return _ingressClassData.ContainsKey(spec.IngressClassName);
+            }
+        }
+
+        return _isDefaultController;
     }
 
     private NamespaceCache Namespace(string key)
