@@ -14,6 +14,13 @@ internal class YarpConfigContext
 
     public List<ClusterConfig> BuildClusterConfig()
     {
-        return ClusterTransfers.Values.Select(c => new ClusterConfig() { Destinations = c.Destinations, ClusterId = c.ClusterId }).ToList();
+        return ClusterTransfers.Values.Select(c => new ClusterConfig() {
+            Destinations = c.Destinations,
+            ClusterId = c.ClusterId,
+            HealthCheck = c.HealthCheck,
+            LoadBalancingPolicy = c.LoadBalancingPolicy,
+            SessionAffinity = c.SessionAffinity,
+            HttpClient = c.HttpClientConfig
+        }).ToList();
     }
 }
