@@ -6,10 +6,13 @@ using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Yarp.ReverseProxy.Configuration;
+#if NET6_0_OR_GREATER
+using Yarp.ReverseProxy.Delegation;
+#endif
+using Yarp.ReverseProxy.Forwarder;
 using Yarp.ReverseProxy.Health;
 using Yarp.ReverseProxy.LoadBalancing;
 using Yarp.ReverseProxy.Model;
-using Yarp.ReverseProxy.Forwarder;
 using Yarp.ReverseProxy.Routing;
 using Yarp.ReverseProxy.SessionAffinity;
 using Yarp.ReverseProxy.Transforms;
@@ -50,6 +53,7 @@ internal static class IReverseProxyBuilderExtensions
     public static IReverseProxyBuilder AddConfigManager(this IReverseProxyBuilder builder)
     {
         builder.Services.TryAddSingleton<ProxyConfigManager>();
+        builder.Services.TryAddSingleton<IProxyStateLookup>(sp => sp.GetRequiredService<ProxyConfigManager>());
         return builder;
     }
 
@@ -110,6 +114,16 @@ internal static class IReverseProxyBuilderExtensions
     public static IReverseProxyBuilder AddPassiveHealthCheck(this IReverseProxyBuilder builder)
     {
         builder.Services.AddSingleton<IPassiveHealthCheckPolicy, TransportFailureRateHealthPolicy>();
+        return builder;
+    }
+
+    public static IReverseProxyBuilder AddHttpSysDelegation(this IReverseProxyBuilder builder)
+    {
+#if NET6_0_OR_GREATER
+        builder.Services.AddSingleton<HttpSysDelegator>();
+        builder.Services.AddSingleton<IClusterChangeListener>(p => p.GetRequiredService<HttpSysDelegator>());
+#endif
+
         return builder;
     }
 }
