@@ -27,7 +27,11 @@ internal sealed class AffinitizeTransform : ResponseTransform
         var proxyFeature = context.HttpContext.GetReverseProxyFeature();
         var options = proxyFeature.Cluster.Config.SessionAffinity;
         // The transform should only be added to routes that have affinity enabled.
-        Debug.Assert(options?.Enabled ?? true, "Session affinity is not enabled");
+        // However, the cluster can be re-assigned dynamically.
+        if (options == null || !options.Enabled.GetValueOrDefault())
+        {
+            return default;
+        }
         var selectedDestination = proxyFeature.ProxiedDestination!;
         _sessionAffinityPolicy.AffinitizeResponse(context.HttpContext, proxyFeature.Route.Cluster!, options!, selectedDestination);
         return default;
