@@ -5,33 +5,55 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Text;
 using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
+using FluentAssertions.Json;
 using k8s;
 using k8s.Models;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Microsoft.Kubernetes;
+using Moq;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Linq;
 using Xunit;
+using Yarp.Kubernetes.Controller;
 using Yarp.Kubernetes.Controller.Caching;
 using Yarp.Kubernetes.Controller.Converters;
 using Yarp.Kubernetes.Controller.Services;
+using JsonSerializer = System.Text.Json.JsonSerializer;
 
 namespace IngressController.Tests;
 
 public class IngressControllerTests
 {
+    public IngressControllerTests()
+    {
+        JsonConvert.DefaultSettings = () => new JsonSerializerSettings() {
+            NullValueHandling = NullValueHandling.Ignore,
+            Converters = {new StringEnumConverter()}
+        };
+    }
+
     [Theory]
     [InlineData("basic-ingress")]
     [InlineData("multiple-endpoints-ports")]
     [InlineData("https")]
     [InlineData("exact-match")]
+    [InlineData("annotations")]
     [InlineData("mapped-port")]
     [InlineData("hostname-routing")]
     [InlineData("multiple-ingresses")]
     [InlineData("multiple-ingresses-one-svc")]
+    [InlineData("multiple-namespaces")]
+    [InlineData("route-metadata")]
     public async Task ParsingTests(string name)
     {
-        var cache = await GetKubernetesInfo(name).ConfigureAwait(false);
+        var ingressClass = KubeResourceGenerator.CreateIngressClass("yarp", "microsoft.com/ingress-yarp", true);
+        var cache = await GetKubernetesInfo(name, ingressClass).ConfigureAwait(false);
         var configContext = new YarpConfigContext();
         var ingresses = cache.GetIngresses().ToArray();
 
@@ -43,9 +65,9 @@ public class IngressControllerTests
                 YarpParser.ConvertFromKubernetesIngress(ingressContext, configContext);
             }
         }
-
-        VerifyClusters(JsonSerializer.Serialize(configContext.BuildClusterConfig()), name);
-        VerifyRoutes(JsonSerializer.Serialize(configContext.Routes), name);
+        var options = new JsonSerializerOptions { Converters = {new JsonStringEnumConverter()} };
+        VerifyClusters(JsonSerializer.Serialize(configContext.BuildClusterConfig(), options), name);
+        VerifyRoutes(JsonSerializer.Serialize(configContext.Routes, options), name);
     }
 
     private static void VerifyClusters(string clusterJson, string name)
@@ -58,21 +80,59 @@ public class IngressControllerTests
         VerifyJson(routesJson, name, "routes.json");
     }
 
+        private static string StripNullProperties(string json)
+        {
+            using var reader = new JsonTextReader(new StringReader(json));
+            var sb = new StringBuilder();
+            using var sw = new StringWriter(sb);
+            using var writer = new JsonTextWriter(sw);
+            while (reader.Read())
+            {
+                var token = reader.TokenType;
+                var value = reader.Value;
+                if(reader.TokenType == JsonToken.PropertyName)
+                {
+                    reader.Read();
+                    if (reader.TokenType == JsonToken.Null)
+                    {
+                        continue;
+                    }
+                    writer.WriteToken(token, value);
+                }
+                writer.WriteToken(reader.TokenType, reader.Value);
+            }
+
+            return sb.ToString();
+        }
     private static void VerifyJson(string json, string name, string fileName)
     {
         var other = File.ReadAllText(Path.Combine("testassets", name, fileName));
+        json = StripNullProperties(json);
+        other = StripNullProperties(other);
 
-        Assert.True(JToken.DeepEquals(JToken.Parse(json), JToken.Parse(other)));
+        var actual = JToken.Parse(json);
+        var jOther = JToken.Parse(other);
+        actual.Should().BeEquivalentTo(jOther);
     }
 
-    private async Task<ICache> GetKubernetesInfo(string name)
+    private async Task<ICache> GetKubernetesInfo(string name, V1IngressClass ingressClass)
     {
-        var cache = new IngressCache();
+        var mockLogger = new Mock<ILogger<IngressCache>>();
+        var mockOptions = new Mock<IOptions<YarpOptions>>();
+
+        mockOptions.SetupGet(o => o.Value).Returns(new YarpOptions { ControllerClass = "microsoft.com/ingress-yarp" });
+
+        var cache = new IngressCache(mockOptions.Object, mockLogger.Object);
 
         var typeMap = new Dictionary<string, Type>();
         typeMap.Add("networking.k8s.io/v1/Ingress", typeof(V1Ingress));
         typeMap.Add("v1/Service", typeof(V1Service));
         typeMap.Add("v1/Endpoints", typeof(V1Endpoints));
+
+        if (ingressClass != null)
+        {
+            cache.Update(WatchEventType.Added, ingressClass);
+        }
 
         var kubeObjects = await Yaml.LoadAllFromFileAsync(Path.Combine("testassets", name, "ingress.yaml"), typeMap).ConfigureAwait(false);
         foreach (var obj in kubeObjects)
