@@ -9,31 +9,55 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Hosting.Server;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Moq;
 using Xunit;
-using Yarp.Tests.Common;
 using Yarp.ReverseProxy.Configuration;
 using Yarp.ReverseProxy.Configuration.ConfigProvider;
-using Yarp.ReverseProxy.Health;
-using Yarp.ReverseProxy.Model;
 using Yarp.ReverseProxy.Forwarder;
 using Yarp.ReverseProxy.Forwarder.Tests;
+using Yarp.ReverseProxy.Health;
+using Yarp.ReverseProxy.Model;
 using Yarp.ReverseProxy.Routing;
+using Yarp.Tests.Common;
 
 namespace Yarp.ReverseProxy.Management.Tests;
 
 public class ProxyConfigManagerTests
 {
-    private IServiceProvider CreateServices(List<RouteConfig> routes, List<ClusterConfig> clusters, Action<IReverseProxyBuilder> configureProxy = null)
+    private static IServiceProvider CreateServices(List<RouteConfig> routes, List<ClusterConfig> clusters, Action<IReverseProxyBuilder> configureProxy = null)
     {
         var serviceCollection = new ServiceCollection();
         serviceCollection.AddLogging();
         serviceCollection.AddRouting();
         var proxyBuilder = serviceCollection.AddReverseProxy().LoadFromMemory(routes, clusters);
+        serviceCollection.TryAddSingleton(new Mock<IServer>().Object);
+        serviceCollection.TryAddSingleton(new Mock<IWebHostEnvironment>().Object);
+        var activeHealthPolicy = new Mock<IActiveHealthCheckPolicy>();
+        activeHealthPolicy.SetupGet(p => p.Name).Returns("activePolicyA");
+        serviceCollection.AddSingleton(activeHealthPolicy.Object);
+        configureProxy?.Invoke(proxyBuilder);
+        var services = serviceCollection.BuildServiceProvider();
+        var routeBuilder = services.GetRequiredService<ProxyEndpointFactory>();
+        routeBuilder.SetProxyPipeline(context => Task.CompletedTask);
+        return services;
+    }
+
+    private static IServiceProvider CreateServices(IEnumerable<IProxyConfigProvider> configProviders, Action<IReverseProxyBuilder> configureProxy = null)
+    {
+        var serviceCollection = new ServiceCollection();
+        serviceCollection.AddLogging();
+        serviceCollection.AddRouting();
+        var proxyBuilder = serviceCollection.AddReverseProxy();
+        foreach (var configProvider in configProviders)
+        {
+            serviceCollection.AddSingleton(configProvider);
+        }
+        serviceCollection.TryAddSingleton(new Mock<IServer>().Object);
         serviceCollection.TryAddSingleton(new Mock<IWebHostEnvironment>().Object);
         var activeHealthPolicy = new Mock<IActiveHealthCheckPolicy>();
         activeHealthPolicy.SetupGet(p => p.Name).Returns("activePolicyA");
@@ -86,6 +110,20 @@ public class ProxyConfigManagerTests
     }
 
     [Fact]
+    public async Task Lookup_StartsEmpty()
+    {
+        var services = CreateServices(new List<RouteConfig>(), new List<ClusterConfig>());
+        var manager = services.GetRequiredService<ProxyConfigManager>();
+        var lookup = services.GetRequiredService<IProxyStateLookup>();
+        await manager.InitialLoadAsync();
+
+        Assert.Empty(lookup.GetRoutes());
+        Assert.Empty(lookup.GetClusters());
+        Assert.False(lookup.TryGetRoute("route1", out var _));
+        Assert.False(lookup.TryGetCluster("cluster1", out var _));
+    }
+
+    [Fact]
     public async Task GetChangeToken_InitialValue()
     {
         var services = CreateServices(new List<RouteConfig>(), new List<ClusterConfig>());
@@ -121,12 +159,89 @@ public class ProxyConfigManagerTests
         var services = CreateServices(new List<RouteConfig>() { route }, new List<ClusterConfig>() { cluster });
 
         var manager = services.GetRequiredService<ProxyConfigManager>();
+        var lookup = services.GetRequiredService<IProxyStateLookup>();
         var dataSource = await manager.InitialLoadAsync();
 
         Assert.NotNull(dataSource);
         var endpoints = dataSource.Endpoints;
         var endpoint = Assert.Single(endpoints);
         var routeConfig = endpoint.Metadata.GetMetadata<RouteModel>();
+        Assert.NotNull(routeConfig);
+        Assert.Equal("route1", routeConfig.Config.RouteId);
+        Assert.True(lookup.TryGetRoute("route1", out var routeModel));
+        Assert.Equal(route, routeModel.Config);
+        routeModel = Assert.Single(lookup.GetRoutes());
+        Assert.Equal(route, routeModel.Config);
+
+        var clusterState = routeConfig.Cluster;
+        Assert.NotNull(clusterState);
+
+        Assert.Equal("cluster1", clusterState.ClusterId);
+        Assert.NotNull(clusterState.Destinations);
+        Assert.NotNull(clusterState.Model);
+        Assert.NotNull(clusterState.Model.HttpClient);
+        Assert.Same(clusterState, routeConfig.Cluster);
+        Assert.True(lookup.TryGetCluster("cluster1", out clusterState));
+        Assert.Equal(cluster, clusterState.Model.Config);
+        clusterState = Assert.Single(lookup.GetClusters());
+        Assert.Equal(cluster, clusterState.Model.Config);
+
+        var actualDestinations = clusterState.Destinations.Values;
+        var destination = Assert.Single(actualDestinations);
+        Assert.Equal("d1", destination.DestinationId);
+        Assert.NotNull(destination.Model);
+        Assert.Equal(TestAddress, destination.Model.Config.Address);
+    }
+
+    [Fact]
+    public async Task BuildConfig_TwoDistinctConfigs_Works()
+    {
+        const string TestAddress = "https://localhost:123/";
+
+        var cluster1 = new ClusterConfig
+        {
+            ClusterId = "cluster1",
+            Destinations = new Dictionary<string, DestinationConfig>(StringComparer.OrdinalIgnoreCase)
+            {
+                { "d1", new DestinationConfig { Address = TestAddress } }
+            }
+        };
+        var route1 = new RouteConfig
+        {
+            RouteId = "route1",
+            ClusterId = "cluster1",
+            Match = new RouteMatch { Path = "/" }
+        };
+
+        var cluster2 = new ClusterConfig
+        {
+            ClusterId = "cluster2",
+            Destinations = new Dictionary<string, DestinationConfig>(StringComparer.OrdinalIgnoreCase)
+            {
+                { "d2", new DestinationConfig { Address = TestAddress } }
+            }
+        };
+        var route2 = new RouteConfig
+        {
+            RouteId = "route2",
+            ClusterId = "cluster2",
+            Match = new RouteMatch { Path = "/" }
+        };
+
+        var config1 = new InMemoryConfigProvider(new List<RouteConfig>() { route1 }, new List<ClusterConfig>() { cluster1 });
+        var config2 = new InMemoryConfigProvider(new List<RouteConfig>() { route2 }, new List<ClusterConfig>() { cluster2 });
+
+        var services = CreateServices(new[] { config1, config2 });
+
+        var manager = services.GetRequiredService<ProxyConfigManager>();
+        var dataSource = await manager.InitialLoadAsync();
+
+        Assert.NotNull(dataSource);
+        var endpoints = dataSource.Endpoints;
+        Assert.Equal(2, endpoints.Count);
+
+        // The order is unstable because routes are stored in a dictionary.
+        var routeConfig = endpoints.Single(e => string.Equals(e.DisplayName, "route1")).Metadata.GetMetadata<RouteModel>();
         Assert.NotNull(routeConfig);
         Assert.Equal("route1", routeConfig.Config.RouteId);
 
@@ -142,6 +257,112 @@ public class ProxyConfigManagerTests
         var actualDestinations = clusterState.Destinations.Values;
         var destination = Assert.Single(actualDestinations);
         Assert.Equal("d1", destination.DestinationId);
+        Assert.NotNull(destination.Model);
+        Assert.Equal(TestAddress, destination.Model.Config.Address);
+
+        routeConfig = endpoints.Single(e => string.Equals(e.DisplayName, "route2")).Metadata.GetMetadata<RouteModel>();
+        Assert.NotNull(routeConfig);
+        Assert.Equal("route2", routeConfig.Config.RouteId);
+
+        clusterState = routeConfig.Cluster;
+        Assert.NotNull(clusterState);
+
+        Assert.Equal("cluster2", clusterState.ClusterId);
+        Assert.NotNull(clusterState.Destinations);
+        Assert.NotNull(clusterState.Model);
+        Assert.NotNull(clusterState.Model.HttpClient);
+        Assert.Same(clusterState, routeConfig.Cluster);
+
+        actualDestinations = clusterState.Destinations.Values;
+        destination = Assert.Single(actualDestinations);
+        Assert.Equal("d2", destination.DestinationId);
+        Assert.NotNull(destination.Model);
+        Assert.Equal(TestAddress, destination.Model.Config.Address);
+    }
+
+    [Fact]
+    public async Task BuildConfig_TwoOverlappingConfigs_Works()
+    {
+        const string TestAddress = "https://localhost:123/";
+
+        var cluster1 = new ClusterConfig
+        {
+            ClusterId = "cluster1",
+            Destinations = new Dictionary<string, DestinationConfig>(StringComparer.OrdinalIgnoreCase)
+            {
+                { "d1", new DestinationConfig { Address = TestAddress } }
+            }
+        };
+        var cluster2 = new ClusterConfig
+        {
+            ClusterId = "cluster2",
+            Destinations = new Dictionary<string, DestinationConfig>(StringComparer.OrdinalIgnoreCase)
+            {
+                { "d2", new DestinationConfig { Address = TestAddress } }
+            }
+        };
+
+        var route1 = new RouteConfig
+        {
+            RouteId = "route1",
+            ClusterId = "cluster1",
+            Match = new RouteMatch { Path = "/" }
+        };
+        var route2 = new RouteConfig
+        {
+            RouteId = "route2",
+            ClusterId = "cluster2",
+            Match = new RouteMatch { Path = "/" }
+        };
+
+        var config1 = new InMemoryConfigProvider(new List<RouteConfig>() { route2 }, new List<ClusterConfig>() { cluster1 });
+        var config2 = new InMemoryConfigProvider(new List<RouteConfig>() { route1 }, new List<ClusterConfig>() { cluster2 });
+
+        var services = CreateServices(new[] { config1, config2 });
+
+        var manager = services.GetRequiredService<ProxyConfigManager>();
+        var dataSource = await manager.InitialLoadAsync();
+
+        Assert.NotNull(dataSource);
+        var endpoints = dataSource.Endpoints;
+        Assert.Equal(2, endpoints.Count);
+
+        // The order is unstable because routes are stored in a dictionary.
+        var routeConfig = endpoints.Single(e => string.Equals(e.DisplayName, "route1")).Metadata.GetMetadata<RouteModel>();
+        Assert.NotNull(routeConfig);
+        Assert.Equal("route1", routeConfig.Config.RouteId);
+
+        var clusterState = routeConfig.Cluster;
+        Assert.NotNull(clusterState);
+
+        Assert.Equal("cluster1", clusterState.ClusterId);
+        Assert.NotNull(clusterState.Destinations);
+        Assert.NotNull(clusterState.Model);
+        Assert.NotNull(clusterState.Model.HttpClient);
+        Assert.Same(clusterState, routeConfig.Cluster);
+
+        var actualDestinations = clusterState.Destinations.Values;
+        var destination = Assert.Single(actualDestinations);
+        Assert.Equal("d1", destination.DestinationId);
+        Assert.NotNull(destination.Model);
+        Assert.Equal(TestAddress, destination.Model.Config.Address);
+
+        routeConfig = endpoints.Single(e => string.Equals(e.DisplayName, "route2")).Metadata.GetMetadata<RouteModel>();
+        Assert.NotNull(routeConfig);
+        Assert.Equal("route2", routeConfig.Config.RouteId);
+
+        clusterState = routeConfig.Cluster;
+        Assert.NotNull(clusterState);
+
+        Assert.Equal("cluster2", clusterState.ClusterId);
+        Assert.NotNull(clusterState.Destinations);
+        Assert.NotNull(clusterState.Model);
+        Assert.NotNull(clusterState.Model.HttpClient);
+        Assert.Same(clusterState, routeConfig.Cluster);
+
+        actualDestinations = clusterState.Destinations.Values;
+        destination = Assert.Single(actualDestinations);
+        Assert.Equal("d2", destination.DestinationId);
         Assert.NotNull(destination.Model);
         Assert.Equal(TestAddress, destination.Model.Config.Address);
     }
@@ -251,6 +472,57 @@ public class ProxyConfigManagerTests
     }
 
     [Fact]
+    public async Task GetChangeToken_MultipleConfigs_SignalsChange()
+    {
+        var config1 = new InMemoryConfigProvider(new List<RouteConfig>(), new List<ClusterConfig>());
+        var config2 = new InMemoryConfigProvider(new List<RouteConfig>(), new List<ClusterConfig>());
+        var services = CreateServices(new[] { config1, config2 });
+        var configManager = services.GetRequiredService<ProxyConfigManager>();
+        var dataSource = await configManager.InitialLoadAsync();
+        _ = configManager.Endpoints; // Lazily creates endpoints the first time, activates change notifications.
+
+        var signaled1 = new TaskCompletionSource<int>(TaskCreationOptions.RunContinuationsAsynchronously);
+        var signaled2 = new TaskCompletionSource<int>(TaskCreationOptions.RunContinuationsAsynchronously);
+        IReadOnlyList<Endpoint> readEndpoints1 = null;
+        IReadOnlyList<Endpoint> readEndpoints2 = null;
+
+        var changeToken1 = dataSource.GetChangeToken();
+        changeToken1.RegisterChangeCallback(
+            _ =>
+            {
+                readEndpoints1 = dataSource.Endpoints;
+                signaled1.SetResult(1);
+            }, null);
+
+        // updating should signal the current change token
+        Assert.False(signaled1.Task.IsCompleted);
+        config1.Update(new List<RouteConfig>() { new RouteConfig() { RouteId = "r1", Match = new RouteMatch { Path = "/" } } }, new List<ClusterConfig>());
+        await signaled1.Task.DefaultTimeout();
+
+        var changeToken2 = dataSource.GetChangeToken();
+        changeToken2.RegisterChangeCallback(
+            _ =>
+            {
+                readEndpoints2 = dataSource.Endpoints;
+                signaled2.SetResult(1);
+            }, null);
+
+        // updating again should only signal the new change token
+        Assert.False(signaled2.Task.IsCompleted);
+        config2.Update(new List<RouteConfig>() { new RouteConfig() { RouteId = "r2", Match = new RouteMatch { Path = "/" } } }, new List<ClusterConfig>());
+        await signaled2.Task.DefaultTimeout();
+
+        var endpoint = Assert.Single(readEndpoints1);
+        Assert.Equal("r1", endpoint.DisplayName);
+
+        Assert.NotNull(readEndpoints2);
+        Assert.Equal(2, readEndpoints2.Count);
+        // Ordering is unstable due to dictionary storage.
+        readEndpoints2.Single(e => string.Equals(e.DisplayName, "r1"));
+        readEndpoints2.Single(e => string.Equals(e.DisplayName, "r2"));
+    }
+
+    [Fact]
     public async Task ChangeConfig_ActiveHealthCheckIsEnabled_RunInitialCheck()
     {
         var endpoints = new List<RouteConfig>() { new RouteConfig() { RouteId = "r1", ClusterId = "c1", Match = new RouteMatch { Path = "/" } } };
@@ -290,6 +562,55 @@ public class ProxyConfigManagerTests
         await signaled.Task.DefaultTimeout();
 
         Assert.True(activeMonitor.Scheduler.IsScheduled(clusterState));
+    }
+
+    [Fact]
+    public async Task ChangeConfig_DestinationChange_IsReflectedOnRouteConfiguration()
+    {
+        var endpoints = new List<RouteConfig>() { new RouteConfig() { RouteId = "r1", ClusterId = "c1", Match = new RouteMatch { Path = "/" } } };
+        var clusters = new List<ClusterConfig>()
+        {
+            new ClusterConfig
+            {
+                ClusterId = "c1",
+                Destinations = new Dictionary<string, DestinationConfig>(StringComparer.OrdinalIgnoreCase)
+                {
+                    { "d1", new DestinationConfig() { Address = "http://d1" } }
+                }
+            }
+        };
+        var services = CreateServices(endpoints, clusters);
+        var inMemoryConfig = (InMemoryConfigProvider)services.GetRequiredService<IProxyConfigProvider>();
+        var configManager = services.GetRequiredService<ProxyConfigManager>();
+        var dataSource = await configManager.InitialLoadAsync();
+
+        var endpoint = Assert.Single(dataSource.Endpoints);
+        var routeConfig = endpoint.Metadata.GetMetadata<RouteModel>();
+        Assert.Equal("http://d1", Assert.Single(routeConfig.Cluster.Destinations).Value.Model.Config.Address);
+        Assert.Equal(1, routeConfig.Cluster.Revision);
+
+        inMemoryConfig.Update(
+            endpoints,
+            new List<ClusterConfig>()
+            {
+                new ClusterConfig
+                {
+                    ClusterId = "c1",
+                    Destinations = new Dictionary<string, DestinationConfig>(StringComparer.OrdinalIgnoreCase)
+                    {
+                        { "d1", new DestinationConfig() { Address = "http://d1-v2" } }
+                    }
+                }
+            });
+
+        var destinationConfig = Assert.Single(routeConfig.Cluster.Destinations).Value.Model.Config;
+        Assert.Equal("http://d1-v2", destinationConfig.Address);
+
+        Assert.Same(destinationConfig, Assert.Single(routeConfig.Cluster.DestinationsState.AllDestinations).Model.Config);
+        Assert.Same(destinationConfig, Assert.Single(routeConfig.Cluster.Model.Config.Destinations).Value);
+
+        // Destination changes do not affect this property
+        Assert.Equal(1, routeConfig.Cluster.Revision);
     }
 
     [Fact]
@@ -334,6 +655,27 @@ public class ProxyConfigManagerTests
         Assert.Single(agex.InnerExceptions);
         var argex = Assert.IsType<ArgumentException>(agex.InnerExceptions.First());
         Assert.StartsWith($"Route '{routeName}' requires Hosts or Path specified", argex.Message);
+    }
+
+    [Fact]
+    public async Task LoadAsync_MultipleSourcesWithValidationErrors_Throws()
+    {
+        var route1 = new RouteConfig { RouteId = "route1", Match = new RouteMatch { Hosts = null }, ClusterId = "cluster1" };
+        var provider1 = new InMemoryConfigProvider(new List<RouteConfig>() { route1 }, new List<ClusterConfig>());
+        var cluster1 = new ClusterConfig { ClusterId = "cluster1", HttpClient = new HttpClientConfig { MaxConnectionsPerServer = -1 } };
+        var provider2 = new InMemoryConfigProvider(new List<RouteConfig>(), new List<ClusterConfig>() { cluster1 });
+        var services = CreateServices(new[] { provider1, provider2 });
+        var configManager = services.GetRequiredService<ProxyConfigManager>();
+
+        var ioEx = await Assert.ThrowsAsync<InvalidOperationException>(() => configManager.InitialLoadAsync());
+        Assert.Equal("Unable to load or apply the proxy configuration.", ioEx.Message);
+        var agex = Assert.IsType<AggregateException>(ioEx.InnerException);
+
+        Assert.Equal(2, agex.InnerExceptions.Count);
+        var argex = Assert.IsType<ArgumentException>(agex.InnerExceptions.First());
+        Assert.StartsWith($"Route 'route1' requires Hosts or Path specified", argex.Message);
+        argex = Assert.IsType<ArgumentException>(agex.InnerExceptions.Skip(1).First());
+        Assert.StartsWith($"Max connections per server limit set on the cluster 'cluster1' must be positive.", argex.Message);
     }
 
     [Fact]
